fix: Move Download button
