+++ components/gallery/GalleryItemButton/GalleryItemButton.vue
@@ -6,9 +6,6 @@
 
     <GalleryItemMoreActionBtn
       :nft="nft"
-      :image-url="nftMetadata?.image"
-      :image-data="imageData"
-      :mime-type="nftMimeType"
       :abi="abi"
       data-testid="gallery-item-more-button"
     />
@@ -20,9 +17,8 @@ import GalleryItemMoreActionBtn from './GalleryItemMoreActionBtn.vue'
 import { extractTwitterIdFromDescription } from '@/utils/parse'
 
 const { $i18n } = useNuxtApp()
-const imageData = ref()
 
-const { getNft: nft, getNftMetadata: nftMetadata, getNftMimeType: nftMimeType, getAbi: abi } = storeToRefs(useNftStore())
+const { getNft: nft, getAbi: abi } = storeToRefs(useNftStore())
 
 const customSharingContent = computed(() => {
   const twitterId = nft.value?.meta?.description
@@ -31,8 +27,6 @@ const customSharingContent = computed(() => {
 
   return twitterId ? $i18n.t('sharing.nftWithArtist', [twitterId]) : ''
 })
-
-onKodahashRenderCompleted(({ payload }) => imageData.value = payload.image)
 </script>
 
 <style scoped lang="scss">
